create a new ViewSet for PHR auth APIs
